backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.9.0
pydantic-settings==2.5.0
pydantic[email]==2.9.0
bcrypt==4.2.0
PyJWT==2.9.0
httpx==0.27.0
stripe==10.12.0
redis==5.1.0
python-multipart==0.0.12
